refactor(gui): retain core surfaces within budget
