+++ framework/gui/src/renderer/src/projects-panel/index.tsx
@@ -9,6 +9,7 @@ import type {
 import * as capability from '@kungfu-tech/api/capability';
 import { mono, panelStyle } from '@kungfu-tech/kfx';
 import React from 'react';
+import { ProjectWorkControlView } from '../../../../../../extensions/work-dashboard/src/view/index';
 import { isResettableRuntimeFailure } from '../../../runtime-recovery-contract';
 import {
   RUNTIME_BACKUP_RESET_CHANNEL,
@@ -20,6 +21,7 @@ import {
   WORKSPACE_START_CONTINUATION_CHANNEL,
 } from '../../../sandbox/channels';
 import { createAgentSessionProxy } from '../agent-session-proxy';
+import { AgentWorkLabPanel } from '../agent-work-lab';
 import { guiKungfuCliArgs } from '../runtime';
 
 export function openRendererProjects() {
@@ -924,4 +926,123 @@ export class KfxErrorBoundary extends React.Component<
     }
     return this.props.children;
   }
+}
+
+export type CoreSurfaceId = 'projects' | 'agent-work-lab' | 'core-work';
+
+export function useRetainedCoreSurfaces({
+  projectsOpen,
+  labOpen,
+  coreWorkOpen,
+}: {
+  projectsOpen: boolean;
+  labOpen: boolean;
+  coreWorkOpen: boolean;
+}): ReadonlySet<CoreSurfaceId> {
+  const [retained, setRetained] = React.useState<ReadonlySet<CoreSurfaceId>>(
+    () =>
+      new Set([
+        ...(projectsOpen ? (['projects'] as const) : []),
+        ...(labOpen ? (['agent-work-lab'] as const) : []),
+        ...(coreWorkOpen ? (['core-work'] as const) : []),
+      ]),
+  );
+  React.useEffect(() => {
+    const visible = projectsOpen
+      ? 'projects'
+      : labOpen
+        ? 'agent-work-lab'
+        : coreWorkOpen
+          ? 'core-work'
+          : undefined;
+    if (!visible) return;
+    setRetained((current) =>
+      current.has(visible) ? current : new Set([...current, visible]),
+    );
+  }, [coreWorkOpen, labOpen, projectsOpen]);
+  return retained;
+}
+
+type ProjectsPanelProps = React.ComponentProps<typeof ProjectsPanel>;
+type LabPanelProps = React.ComponentProps<typeof AgentWorkLabPanel>;
+type WorkPanelProps = React.ComponentProps<typeof ProjectWorkControlView>;
+
+export function RetainedCoreSurfaceStack({
+  visible,
+  retained,
+  projects,
+  focusedPath,
+  onCatalog,
+  onOpenProject,
+  onOpenExistingProject,
+  onRestoreProject,
+  lab,
+  startup,
+  onOpenWork,
+  onOpenLabExistingProject,
+  onOpenStarterProject,
+  shell,
+}: {
+  visible?: CoreSurfaceId;
+  retained: ReadonlySet<CoreSurfaceId>;
+  projects: ProjectsPanelProps['projects'] & WorkPanelProps['projects'];
+  focusedPath: ProjectsPanelProps['focusedPath'];
+  onCatalog: ProjectsPanelProps['onCatalog'];
+  onOpenProject: ProjectsPanelProps['onOpenProject'];
+  onOpenExistingProject: ProjectsPanelProps['onOpenExistingProject'];
+  onRestoreProject: ProjectsPanelProps['onRestoreProject'];
+  lab: LabPanelProps['lab'];
+  startup: LabPanelProps['startup'];
+  onOpenWork: LabPanelProps['onOpenWork'];
+  onOpenLabExistingProject: LabPanelProps['onOpenExistingProject'];
+  onOpenStarterProject: LabPanelProps['onOpenStarterProject'];
+  shell: WorkPanelProps['shell'];
+}) {
+  return (
+    <>
+      {visible === 'projects' || retained.has('projects') ? (
+        <div
+          style={{
+            display: visible === 'projects' ? 'block' : 'none',
+            height: '100%',
+          }}
+        >
+          <ProjectsPanel
+            projects={projects}
+            focusedPath={focusedPath}
+            onCatalog={onCatalog}
+            onOpenProject={onOpenProject}
+            onOpenExistingProject={onOpenExistingProject}
+            onRestoreProject={onRestoreProject}
+          />
+        </div>
+      ) : null}
+      {visible === 'agent-work-lab' || retained.has('agent-work-lab') ? (
+        <div
+          style={{
+            display: visible === 'agent-work-lab' ? 'block' : 'none',
+            height: '100%',
+          }}
+        >
+          <AgentWorkLabPanel
+            lab={lab}
+            startup={startup}
+            onOpenWork={onOpenWork}
+            onOpenExistingProject={onOpenLabExistingProject}
+            onOpenStarterProject={onOpenStarterProject}
+          />
+        </div>
+      ) : null}
+      {visible === 'core-work' || retained.has('core-work') ? (
+        <div
+          style={{
+            display: visible === 'core-work' ? 'block' : 'none',
+            height: '100%',
+          }}
+        >
+          <ProjectWorkControlView projects={projects} shell={shell} />
+        </div>
+      ) : null}
+    </>
+  );
 }
